% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangevoting.R
\name{majority}
\alias{majority}
\title{Majority system calculator}
\usage{
majority(counter)
}
\arguments{
\item{counter}{"Counter of total parliament seats for each party in previously analyzed costituencies"}
}
\value{
Updated counter of total parliament seats
}
\description{
Reads a csv file containing votes and computes the winner of that costituency in a majority system
}
\examples{
counter = c(1,0,3,2,0,5) #suppose 6 parties are running in the election
counter = majority(counter)
}
