% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangevoting.R
\name{proportional}
\alias{proportional}
\title{Proportional system calculator}
\usage{
proportional(seats, counter)
}
\arguments{
\item{seats}{"Number of seats assigned to the analyzed costituency"}

\item{counter}{"Named vector of total parliament seats won by each party in previously analyzed costituencies"}
}
\value{
Updated counter of parliament seats for each party
}
\description{
Reads a csv file containing votes for each party in a costituency and assigns the corresponding number of seats
}
\examples{
seats = 10
counter = c(1,2,3,0,4,1) #in the used dataset, 6 parties are running in the election
counter = proportional(seats,counter)

}
