% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangevoting.R
\name{randomvote}
\alias{randomvote}
\title{Random vote generator}
\usage{
randomvote(N, candidates, parties)
}
\arguments{
\item{N}{"Number of voters"}

\item{candidates}{"Names of candidates"}

\item{parties}{"Names of parties associated to candidates"}
}
\value{
The data frame containing randomly generated votes
}
\description{
Generates a matrix of random votes from a uniform distribution and writes a csv file
}
\examples{
N = 100
candidates = c('A','B','C')
parties = c('right','center','left')
randomvote(N,candidates,parties)

}
